requirements.txt
z3-solver
discord.py
peewee
